requirements.txt
pandas==1.4.1
numpy==1.22.2
matplotlib==3.5.1
osmread==0.2
scikit-learn==1.0.2
networkx==2.6.3
PyYAML==6.0
